+++ Specs/Scene/ModelExperimental/ModelExperimentalSpec.js
@@ -58,7 +58,11 @@ describe(
       "./Data/Models/GltfLoader/MorphPrimitivesTest/glTF/MorphPrimitivesTest.gltf";
     const animatedTriangleUrl =
       "./Data/Models/GltfLoader/AnimatedTriangle/glTF/AnimatedTriangle.gltf";
-
+    const animatedTriangleOffset = new HeadingPitchRange(
+      CesiumMath.PI / 2.0,
+      0,
+      2.0
+    );
     const pointCloudUrl =
       "./Data/Models/GltfLoader/PointCloudWithRGBColors/glTF-Binary/PointCloudWithRGBColors.glb";
 
@@ -74,10 +78,12 @@ describe(
 
     afterAll(function () {
       scene.destroyForSpecs();
+      sceneWithWebgl2.destroyForSpecs();
     });
 
     afterEach(function () {
       scene.primitives.removeAll();
+      sceneWithWebgl2.primitives.removeAll();
       ResourceCache.clearForSpecs();
     });
 
@@ -137,7 +143,7 @@ describe(
       const modelHasIndices = defaultValue(options.hasIndices, true);
       const targetScene = defaultValue(options.scene, scene);
 
-      const commandList = scene.frameState;
+      const commandList = targetScene.frameState;
       const commandCounts = [];
       let i, command;
 
@@ -333,12 +339,16 @@ describe(
       return loadAndZoomToModelExperimental(
         {
           gltf: animatedTriangleUrl,
+          offset: animatedTriangleOffset,
         },
         scene
       ).then(function (model) {
         const animationCollection = model.activeAnimations;
         expect(animationCollection).toBeDefined();
         expect(animationCollection.length).toBe(0);
+
+        // Move camera so that the triangle is in view.
+        scene.camera.moveDown(0.5);
         verifyRender(model, true, {
           zoomToModel: false,
         });
@@ -349,9 +359,13 @@ describe(
       return loadAndZoomToModelExperimental(
         {
           gltf: animatedTriangleUrl,
+          offset: animatedTriangleOffset,
         },
         scene
       ).then(function (model) {
+        // Move camera so that the triangle is in view.
+        scene.camera.moveDown(0.5);
+
         // The model rotates such that it leaves the view of the camera
         // halfway into its animation.
         const startTime = JulianDate.fromDate(
@@ -505,15 +519,46 @@ describe(
       });
     });
 
-    it("debugWireframe works", function () {
+    it("debugWireframe works for WebGL1 if enableDebugWireframe is true", function () {
       const resource = Resource.createIfNeeded(boxTexturedGlbUrl);
       const loadPromise = resource.fetchArrayBuffer();
       return loadPromise.then(function (buffer) {
         return loadAndZoomToModelExperimental(
-          { gltf: new Uint8Array(buffer) },
+          { gltf: new Uint8Array(buffer), enableDebugWireframe: true },
           scene
         ).then(function (model) {
           verifyDebugWireframe(model, PrimitiveType.TRIANGLES);
+        });
+      });
+    });
+
+    it("debugWireframe does nothing in WebGL1 if enableDebugWireframe is false", function () {
+      const resource = Resource.createIfNeeded(boxTexturedGlbUrl);
+      const loadPromise = resource.fetchArrayBuffer();
+      return loadPromise.then(function (buffer) {
+        return loadAndZoomToModelExperimental(
+          { gltf: new Uint8Array(buffer), enableDebugWireframe: false },
+          scene
+        ).then(function (model) {
+          const commandList = scene.frameState;
+          const commandCounts = [];
+          let i, command;
+          scene.renderForSpecs();
+          for (i = 0; i < commandList.length; i++) {
+            command = commandList[i];
+            expect(command.primitiveType).toBe(PrimitiveType.TRIANGLES);
+            commandCounts.push(command.count);
+          }
+
+          model.debugWireframe = true;
+          expect(model._drawCommandsBuilt).toBe(false);
+
+          scene.renderForSpecs();
+          for (i = 0; i < commandList.length; i++) {
+            command = commandList[i];
+            expect(command.primitiveType).toBe(PrimitiveType.TRIANGLES);
+            expect(command.count).toEqual(commandCounts[i]);
+          }
         });
       });
     });
@@ -538,7 +583,7 @@ describe(
 
     it("debugWireframe works for model without indices", function () {
       return loadAndZoomToModelExperimental(
-        { gltf: triangleWithoutIndicesUrl },
+        { gltf: triangleWithoutIndicesUrl, enableDebugWireframe: true },
         scene
       ).then(function (model) {
         verifyDebugWireframe(model, PrimitiveType.TRIANGLES, {
@@ -549,7 +594,7 @@ describe(
 
     it("debugWireframe works for model with triangle strip", function () {
       return loadAndZoomToModelExperimental(
-        { gltf: triangleStripUrl },
+        { gltf: triangleStripUrl, enableDebugWireframe: true },
         scene
       ).then(function (model) {
         verifyDebugWireframe(model, PrimitiveType.TRIANGLE_STRIP);
@@ -558,7 +603,7 @@ describe(
 
     it("debugWireframe works for model with triangle fan", function () {
       return loadAndZoomToModelExperimental(
-        { gltf: triangleFanUrl },
+        { gltf: triangleFanUrl, enableDebugWireframe: true },
         scene
       ).then(function (model) {
         verifyDebugWireframe(model, PrimitiveType.TRIANGLE_FAN);
@@ -567,7 +612,7 @@ describe(
 
     it("debugWireframe ignores points", function () {
       return loadAndZoomToModelExperimental(
-        { gltf: pointCloudUrl },
+        { gltf: pointCloudUrl, enableDebugWireframe: true },
         scene
       ).then(function (model) {
         const commandList = scene.frameState;
@@ -628,15 +673,16 @@ describe(
       });
     });
 
-    // see https://github.com/CesiumGS/cesium/pull/10115
-    xit("renders model with style", function () {
+    it("renders model with style", function () {
       let model;
       let style;
       return loadAndZoomToModelExperimental({ gltf: buildingsMetadata }, scene)
         .then(function (result) {
           model = result;
           // Renders without style.
-          verifyRender(model, true);
+          verifyRender(model, true, {
+            zoomToModel: false,
+          });
 
           // Renders with opaque style.
           style = new Cesium3DTileStyle({
@@ -648,7 +694,9 @@ describe(
         })
         .then(function () {
           model.style = style;
-          verifyRender(model, true);
+          verifyRender(model, true, {
+            zoomToModel: false,
+          });
 
           // Renders with translucent style.
           style = new Cesium3DTileStyle({
@@ -660,7 +708,9 @@ describe(
         })
         .then(function () {
           model.style = style;
-          verifyRender(model, true);
+          verifyRender(model, true, {
+            zoomToModel: false,
+          });
 
           // Does not render when style disables show.
           style = new Cesium3DTileStyle({
@@ -672,11 +722,15 @@ describe(
         })
         .then(function () {
           model.style = style;
-          verifyRender(model, false);
+          verifyRender(model, false, {
+            zoomToModel: false,
+          });
 
           // Render when style is removed.
           model.style = undefined;
-          verifyRender(model, true);
+          verifyRender(model, true, {
+            zoomToModel: false,
+          });
         });
     });