+++ CHANGES.md
@@ -1,5 +1,19 @@
 # Change Log
 
+### 1.94 - 2022-06-01
+
+##### Breaking Changes :mega:
+
+- Models and tilesets rendered with `ModelExperimental` must set `enableDebugWireframe` to true in order for `debugWireframe` to work in WebGL1. [#10344](https://github.com/CesiumGS/cesium/pull/10344)
+
+##### Additions :tada:
+
+- Added `IndexDatatype.fromTypedArray`. [#10350](https://github.com/CesiumGS/cesium/pull/10350)
+
+##### Fixes :wrench:
+
+- Fixed the inaccurate computation of bounding spheres for `ModelExperimental`. [#10339](https://github.com/CesiumGS/cesium/pull/10339/)
+
 ### 1.93 - 2022-05-02
 
 ##### Breaking Changes :mega: